image_api/image_requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
gdown